+++ tests/testthat/apis/api.osf.io/v2/preprints/khbvy.R
@@ -1,6 +1,6 @@
 structure(list(method = "GET", url = "https://api.osf.io/v2/preprints/khbvy/", 
     status_code = 200L, headers = structure(list(server = "nginx", 
-        date = "Wed, 27 May 2026 11:09:23 GMT", `content-type` = "application/vnd.api+json; charset=utf-8", 
+        date = "Sun, 31 May 2026 12:26:31 GMT", `content-type` = "application/vnd.api+json; charset=utf-8", 
         vary = "Accept-Encoding", vary = "Accept, Cookie, origin", 
         allow = "GET, PUT, PATCH, DELETE, HEAD, OPTIONS", `x-content-type-options` = "nosniff", 
         `referrer-policy` = "same-origin", `cross-origin-opener-policy` = "same-origin", 
@@ -512,5 +512,5 @@ structure(list(method = "GET", url = "https://api.osf.io/v2/preprints/khbvy/",
     0x6d, 0x65, 0x74, 0x61, 0x22, 0x3a, 0x7b, 0x22, 0x76, 0x65, 
     0x72, 0x73, 0x69, 0x6f, 0x6e, 0x22, 0x3a, 0x22, 0x32, 0x2e, 
     0x30, 0x22, 0x7d, 0x7d)), timing = c(redirect = 0, namelookup = 0, 
-    connect = 0, pretransfer = 0.000154, starttransfer = 0.524003, 
-    total = 0.524518), cache = new.env(parent = emptyenv())), class = "httr2_response")
+    connect = 0, pretransfer = 0.000153, starttransfer = 2.639605, 
+    total = 2.640027), cache = new.env(parent = emptyenv())), class = "httr2_response")